+++ osp_scraper/settings/rq.py
