+++ osp_scraper/spidermiddlewares.py
