+++ osp_scraper/spiders/CustomSpider.py
