+++ scrapy.cfg
@@ -4,7 +4,7 @@
 # https://scrapyd.readthedocs.org/en/latest/deploy.html
 
 [settings]
-default = syllascrape.settings.scrapy
+default = osp_scraper.settings.scrapy
 
 [deploy]
-project = syllascrape
+project = osp_scraper
